Save point.  Write commit data out to disk.
This is probably enough for one night.  Tomorrow I'll remove the
hard coded paths, and add a cli interface for storing and retrieving
info.  After that's done I should have a solid enough grasp of this
to figure out a practical (eg not over-normalised) approach for
storing the tree, license, and commit data in PG instead of on disk.

When that's hanging together decently... I think we have the basis
for the new data model to use for DBHub.io. :)