Stub API server for dio to connect to.
I'm starting to get a reasonable understanding of what's needed from
each piece, so making an API server for dio to communicate with seems
like a good next step.

* The server side dio code should be moved into this
* A REST client added to dio then hooked up to this API

Once that's in place, update the API to use PG and Minio backend,
which should give an end-to-end understanding of things.