test: add --remote option, to test against a real DB4S end point
By default (without this option), the test spin up a local "mock"
server to test the HTTPS API calls.

With the --remote option however, the local mock server isn't
started and the HTTPS API calls are done using the url specified.

  $ go test --check.v --remote https://docker-dev.dbhub.io:5550

This is because these tests are likely to be useful for the
development of the DB4S end point code too, as they should help
catch any accidental API breakage.